test(cli): fail closed at the Corsa resolver boundary
